# openapi.yaml
openapi: 3.1.0
info:
  title: ClinicalTrials.gov Query → Visualization Agent API
  version: 1.0.0
  description: |
    Converts a natural-language question about clinical trials into a structured
    visualization specification backed by live ClinicalTrials.gov Data API v2 data.

    Design contract:
    - The LLM only produces a `QueryPlan` (which filters / grouping / chart type).
      Every number in the response is computed by code from real trial records.
    - The agent first decides whether a visualization is even possible (`meta.status`).
      When it is not (`unsupported` / `needs_clarification`), `visualization` is `null`.
    - Counts are exact by default (full pagination, no sampling); `meta.truncated`
      is only `true` when an optional `MAX_STUDIES` cap is configured and exceeded.
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /query:
    post:
      operationId: query
      summary: Interpret a question and return a visualization specification
      description: Plain request/response. For streamed progress use `GET /query/stream`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
            examples:
              phaseDistribution:
                summary: Phase distribution for a condition
                value:
                  query: "How are melanoma trials distributed across phases?"
                  condition: "melanoma"
              timeTrend:
                summary: Time trend with caller-supplied fields
                value:
                  query: "How has the number of trials for this drug changed over time?"
                  drug_name: "Pembrolizumab"
                  start_year: 2015
      responses:
        "200":
          description: A visualization specification (or a no-viz result when `meta.status` != `ok`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          description: Invalid request (failed schema validation).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "502":
          description: Upstream ClinicalTrials.gov API error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"

  /query/stream:
    get:
      operationId: queryStream
      summary: Same as POST /query, streamed over Server-Sent Events
      description: |
        Streams the run as SSE so a client can show the agent's interpretation and live
        fetch progress before the final result. Uses query-string params (so `EventSource`
        can connect). Event sequence:
        `status` → `interpretation` → `fetch_progress`* → `result` → `done`
        (or `error`). The `result` event's `data` is an `AgentResponse`.
      parameters:
        - { in: query, name: query, required: true, schema: { type: string, minLength: 1, maxLength: 2000 } }
        - { in: query, name: drug_name, schema: { type: string, maxLength: 200 } }
        - { in: query, name: condition, schema: { type: string, maxLength: 200 } }
        - { in: query, name: sponsor, schema: { type: string, maxLength: 200 } }
        - { in: query, name: country, schema: { type: string, maxLength: 200 } }
        - { in: query, name: study_type, schema: { type: string, maxLength: 50 } }
        - { in: query, name: start_year, schema: { type: integer, minimum: 1900, maximum: 2100 } }
        - { in: query, name: end_year, schema: { type: integer, minimum: 1900, maximum: 2100 } }
        - in: query
          name: status
          description: Comma-separated overall statuses (e.g. `RECRUITING,COMPLETED`).
          schema: { type: string }
        - in: query
          name: trial_phase
          description: Comma-separated phases.
          schema: { type: string }
        - in: query
          name: visualization_type
          schema: { $ref: "#/components/schemas/VisualizationType" }
      responses:
        "200":
          description: An SSE stream (`text/event-stream`).
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE frames. Each frame is `event: <name>` + `data: <json>`.
                  Event payloads:
                  - status:         { type, message }
                  - interpretation: { type, interpreter, queryPlan, notes?, confidence, warnings? }
                  - fetch_progress: { type, fetched, total, page }
                  - result:         { type, response: AgentResponse }
                  - error:          { error, message?, details? }

  /health:
    get:
      operationId: health
      summary: Liveness probe + active interpreter
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status, interpreter]
                properties:
                  status: { type: string, enum: [ok] }
                  interpreter: { $ref: "#/components/schemas/Interpreter" }

components:
  schemas:
    # ---- Enums ----------------------------------------------------------------
    VisualizationType:
      type: string
      enum: [bar_chart, grouped_bar_chart, time_series, scatter_plot, histogram, network_graph]
    TrialPhase:
      type: string
      enum: [EARLY_PHASE1, PHASE1, PHASE2, PHASE3, PHASE4, NA]
    GroupByDimension:
      type: string
      enum: [phase, year, country, sponsor, sponsor_class, intervention_type, condition, status, study_type]
    PlanStatus:
      type: string
      enum: [ok, needs_clarification, unsupported]
    Interpreter:
      type: string
      enum: [llm, fallback-rules]
    SortDirection:
      type: string
      enum: [asc, desc]

    # ---- Request --------------------------------------------------------------
    QueryRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
          description: Natural-language question about clinical trials.
        drug_name: { type: string, maxLength: 200, description: "Drug/intervention. Overrides inference." }
        condition: { type: string, maxLength: 200, description: "Disease/condition. Overrides inference." }
        sponsor: { type: string, maxLength: 200 }
        country: { type: string, maxLength: 200 }
        trial_phase:
          description: One phase or an array of phases.
          oneOf:
            - $ref: "#/components/schemas/TrialPhase"
            - { type: array, items: { $ref: "#/components/schemas/TrialPhase" } }
        status:
          type: array
          items: { type: string, maxLength: 50 }
          description: Overall recruitment statuses, e.g. ["RECRUITING","COMPLETED"].
        study_type: { type: string, maxLength: 50, description: "INTERVENTIONAL | OBSERVATIONAL | EXPANDED_ACCESS" }
        start_year: { type: integer, minimum: 1900, maximum: 2100 }
        end_year: { type: integer, minimum: 1900, maximum: 2100 }
        visualization_type:
          allOf: [{ $ref: "#/components/schemas/VisualizationType" }]
          description: Force a chart type, bypassing the agent's choice.

    # ---- Response -------------------------------------------------------------
    AgentResponse:
      type: object
      required: [visualization, meta]
      properties:
        visualization:
          description: Null when meta.status != "ok" (no chart produced; see meta.notes).
          oneOf:
            - $ref: "#/components/schemas/VisualizationSpec"
            - type: "null"
        meta: { $ref: "#/components/schemas/ResponseMeta" }
        references:
          type: array
          description: Deduplicated source cards for every trial cited by the visualization.
          items: { $ref: "#/components/schemas/Reference" }
        trace: { $ref: "#/components/schemas/ResponseTrace" }

    VisualizationSpec:
      type: object
      required: [type, title, encoding]
      properties:
        type: { $ref: "#/components/schemas/VisualizationType" }
        title: { type: string }
        encoding:
          type: object
          description: Maps fields to visual channels (x/y/series, or nodes/edges for networks).
          additionalProperties:
            type: object
            properties:
              field: { type: string }
              type: { type: string, description: "ordinal | quantitative | temporal | nominal" }
              label: { type: string }
        data:
          type: array
          description: Tabular rows for chart types (omitted for network_graph).
          items: { $ref: "#/components/schemas/VizDatum" }
        nodes:
          type: array
          description: network_graph only.
          items: { $ref: "#/components/schemas/NetworkNode" }
        edges:
          type: array
          description: network_graph only.
          items: { $ref: "#/components/schemas/NetworkEdge" }

    VizDatum:
      type: object
      description: |
        One renderable row; shape varies by chart type (flat). Common fields shown;
        e.g. bar_chart={category,trial_count}, time_series={year,trial_count},
        grouped_bar_chart={group,series,trial_count}, histogram={bucket,bucket_start,bucket_end,trial_count},
        scatter_plot={x,y,nct_id,label}.
      additionalProperties: true
      properties:
        category: { type: string }
        group: { type: string }
        series: { type: string }
        year: { type: integer }
        bucket: { type: string }
        bucket_start: { type: number }
        bucket_end: { type: number }
        x: { type: number }
        y: { type: number }
        nct_id: { type: string }
        label: { type: string }
        trial_count: { type: integer }
        citations:
          type: array
          items: { $ref: "#/components/schemas/Citation" }

    NetworkNode:
      type: object
      required: [id, label, group, weight]
      properties:
        id: { type: string }
        label: { type: string }
        group: { type: string, description: "entity kind, e.g. sponsor | drug | condition" }
        weight: { type: integer, description: "number of trials the node participates in" }

    NetworkEdge:
      type: object
      required: [source, target, weight]
      properties:
        source: { type: string }
        target: { type: string }
        weight: { type: integer, description: "number of trials connecting the two nodes" }
        citations:
          type: array
          items: { $ref: "#/components/schemas/Citation" }

    Citation:
      type: object
      required: [nct_id, excerpt, url]
      properties:
        nct_id: { type: string, example: "NCT01234567" }
        excerpt: { type: string, description: "exact text excerpt from the API record" }
        url: { type: string, format: uri }

    Reference:
      type: object
      required: [nct_id, title, url]
      properties:
        nct_id: { type: string }
        title: { type: string }
        officialTitle: { type: [string, "null"] }
        sponsor: { type: [string, "null"] }
        investigator: { type: [string, "null"] }
        startDate: { type: [string, "null"] }
        url: { type: string, format: uri }

    ResponseMeta:
      type: object
      required: [source, apiVersion, filters, totalMatchingTrials, analyzedTrials, truncated, confidence, status, interpreter, apiRequests]
      properties:
        source: { type: string, enum: [clinicaltrials.gov] }
        apiVersion: { type: string, enum: [v2] }
        filters:
          type: object
          description: Filters actually applied (the QueryPlan's filters).
          additionalProperties: true
        totalMatchingTrials: { type: integer, description: "API-reported total match count" }
        analyzedTrials: { type: integer, description: "records fetched & aggregated" }
        truncated: { type: boolean, description: "true only if a MAX_STUDIES cap was exceeded (sampled)" }
        groupBy: { $ref: "#/components/schemas/GroupByDimension" }
        units: { type: string }
        sort: { type: string }
        notes: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        status: { $ref: "#/components/schemas/PlanStatus" }
        interpreter: { $ref: "#/components/schemas/Interpreter" }
        apiRequests:
          type: array
          items: { type: string, format: uri }
          description: Exact ClinicalTrials.gov request URLs issued (traceability).
        warnings:
          type: array
          items: { type: string }

    QueryPlan:
      type: object
      description: The structured plan the agent inferred (also surfaced in `trace`).
      required: [status, filters, title, confidence]
      properties:
        status: { $ref: "#/components/schemas/PlanStatus" }
        visualizationType: { $ref: "#/components/schemas/VisualizationType" }
        filters:
          type: object
          additionalProperties: true
        groupBy: { $ref: "#/components/schemas/GroupByDimension" }
        sort: { $ref: "#/components/schemas/SortDirection" }
        topN: { type: integer, minimum: 1, maximum: 100 }
        comparison:
          type: object
          properties:
            dimension: { type: string, enum: [intervention, condition, sponsor] }
            values: { type: array, items: { type: string }, minItems: 2 }
        network:
          type: object
          properties:
            source: { type: string, enum: [sponsor, drug, condition] }
            target: { type: string, enum: [sponsor, drug, condition] }
        scatter:
          type: object
          properties:
            x: { type: string, enum: [enrollment, start_year] }
            y: { type: string, enum: [enrollment, start_year] }
        histogram:
          type: object
          properties:
            field: { type: string, enum: [enrollment, start_year] }
            bucketSize: { type: number }
        title: { type: string }
        notes: { type: string }
        confidence: { type: number }

    ResponseTrace:
      type: object
      description: Verification trace — what the agent inferred + the raw API I/O.
      required: [queryPlan, interpreter, apiRequests, apiTotalCount, apiResponseSample]
      properties:
        queryPlan: { $ref: "#/components/schemas/QueryPlan" }
        interpreter: { $ref: "#/components/schemas/Interpreter" }
        apiRequests:
          type: array
          items: { type: string, format: uri }
        apiTotalCount: { type: integer }
        apiResponseSample:
          type: array
          description: First few RAW, unmodified API study records.
          items: { type: object, additionalProperties: true }

    # ---- Errors ---------------------------------------------------------------
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error: { type: string, enum: [invalid_request] }
        details:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
    UpstreamError:
      type: object
      required: [error, message]
      properties:
        error: { type: string, enum: [clinicaltrials_api_error] }
        message: { type: string }
    InternalError:
      type: object
      required: [error, message]
      properties:
        error: { type: string, enum: [internal_error] }
        message: { type: string }